requirements.txt
flask
python-dotenv
requests
pyjwt
nicegui
aiohttp